Fix examples and linkme shenanigans
The distributed runtime for Balter is dependent on `linkme`, and
specifically using it from the `#[scenario]` procedural macro.
Initially I thought this would be fine to re-export via Balter, but
after some experimentation I learned that dependents of Balter
will have to depend on `linkme` explicitly.

Additionally tried to split up the examples directory to test this
separation between with/without `rt` feature set, but features are
shared in the workspace, so now it is just a cosmetic change.