Fix TPS sampling method
The TPS sampling method used previously is garbage. It relied on each
task completion to act as a timer for collecting the global TPS from the
atomics, which meant that if two tasks completed right next to each
other the second task would have next to no data (and therefore be quite
inaccurate). This lead to instability in all of the various controllers.

The new TPS sampling method is dead simple: each task contains a batch
of the scenarios to be run (which is configured based on the number of
transactions in a scenario). Each task completion gives a fraction of
the total TPS (1/N where N is the number of tasks), which we can use to
give an approximation of the total TPS. This is a far more accurate and
robust method of measuring TPS.