Refactor Scenario runner to be generic rather than Pin<Box<>>
Upon doing some benchmarks with Pin<Box<impl Future>>, I found that it
is _slow_. As in, about an order of magnitude slower than a regular
Future. It was originally a shim to get things working, and replacing
with generics was always the plan. The types were just annoying at the
time.

This gets the generic Future integration done, where Pin<Box<T>> is only
kept around for the outermost Scenario which is used by the distributed
runtime. Unfortunately, the distributed runtime functionality is broken
and will have a followup commit to actually fix it.

I haven't done a ton of benchmarking yet, but so far it seems there are
some *major* performance wins with this change which is exciting.