Refactor ConcurrencyController for better stability
The ConcurrencyController has gone through a handful of iterations, the
latest attempts to simplify the operations drastically. It now
calculates the "optimal" task count given the measured TPS ratio to the
goal TPS and the current concurrency. There needs to be some safeguards
(currently limited to increases by 100) since otherwise in wildly
underprovisioned workflows it attempts to go too high too fast.

Additionally, it uses simple slope detection to attempt to determine if
the server is unable to reach the TPS requested; if we keep increasing
the concurrency but the TPS we are hitting stays the same then there is
no point in continuing to increase concurrency. This functionality will
likely require some fine-tuning.