Refactor ErrorRateController for better stability
Update the ErrorRateController (ERC) with the new ConcurrencyController
(CC) logic, as well as various stability improvements. Unfortunately the
ERC is still a bit finicky, and will likely require another revision to
achieve proper stability. However, this brings it in-line with the new
CC and is better than it was.

Still need to figure out a better automated testing strategy, as this
change is arguably only hand-tested via running the various examples.