Refactor of core control loop and sampling mechanisms
This is a significant commit, with a ton of changes across all sorts of
logic. Unfortunately tricky to fix core logic without affecting the rest
of the codebase.

Additionally, this commit is sort-of a WIP commit, as core functionality
is missing (but everything compiles and is shimmed in). Fixing of
functionality will come in smaller commits.