Redo MockService to support mocking latency
With the addition of the LatencyController we needed some way to test
it. The MockService before this commit is fairly unsophisticated, and
has a separate route for each controller type, which limits usability in
composite controller scenarios (ie. error rate + latency at the same
time).

Create a new default route which takes a JSON object to specify
parameters, and adds logic to simulate a linear scaling latency with
TPS. Adds a test for the LatencyController.