Fix handling looping Scenarios
The original design of Balter was to have the framework handle looping
the Scenario. However, upon iterating on the design it felt *way* better
to have a Scenario which was like so:

```
\#[scenario]
async fn foo_scenario() {
    loop {
        ...
    }
}
```

This also makes it easier to set up a Client which is reused per
transaction. So this commit just makes it so that Balter support
Scenarios which loop indefinitely.