Add basic support for noise to integration tests
Based on a bug report, Balter was struggling to achieve 50TPS and was
limiting at a concurrency of 7. Unfortunately recreating this is proving
to be tricky. This commit adds basic support for simulating noise and
shows that in this case Balter behaves correctly.

Unfortunately still not recreating https://github.com/BalterLoadTesting/balter/issues/8